- Correction du temps en double - Ajout du SimulateurTaxis.java

+++ trunk/src/simulateur/Client.java
@@ -5,9 +5,9 @@ import java.awt.geom.Point2D;
 public class Client {
 
 	private Point2D.Double depart, arrivee;
-	private int tempsAttente;
+	private double tempsAttente;
 
-	public Client(Point2D.Double depart, Point2D.Double arrivee, int tempsAttente) {
+	public Client(Point2D.Double depart, Point2D.Double arrivee, double tempsAttente) {
 		this.depart = depart;
 		this.arrivee = arrivee;
 		this.tempsAttente = tempsAttente;
